+++ packages/tide_kit/lib/src/status_bar/tide_status_bar.dart
@@ -22,7 +22,7 @@ class TideStatusBar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final workbenchService = Tide.get<TideWorkbenchService>();
+    final workbenchService = Tide.getIt<TideWorkbenchService>();
     final statusBarState = workbenchService.accessor
         .get<TideWorkbenchLayoutService>()
         .statusBarState;